Use the right printf args for client_request_buffer_max warning

+++ src/cache_cf.cc
@@ -912,9 +912,9 @@ configDoConfigure(void)
     // prevent infinite fetch loops in the request parser
     // due to buffer full but not enough data recived to finish parse
     if (Config.maxRequestBufferSize <= Config.maxRequestHeaderSize) {
-        fatalf("Client request buffer of %d bytes cannot hold a request with %d bytes of headers." \
+        fatalf("Client request buffer of %u bytes cannot hold a request with %u bytes of headers." \
                " Change client_request_buffer_max or request_header_max_size limits.",
-               Config.maxRequestBufferSize, Config.maxRequestHeaderSize);
+               (uint32_t)Config.maxRequestBufferSize, (uint32_t)Config.maxRequestHeaderSize);
     }
 }
